// src/app/models/term.ts
export class Term {
  id: number;
  startTime: string;
  endTime: string;
  group: number;
  module: string;
  optionalSubjectNumber: number;
  numberOfLectures: number;
  numberOfExercises: number;
  numberOfLabExercises: number;
  classroomId: number;
  classroom: string;
  subjectId: number;
  moduleId: number;
  semesterId: number;
  scheduleId: number;
  lecturerId: number;
  slotId: number;
  lecturerFullName: string;
  subject: string;
  weekday: string;
  weekdayId: number;
  isOptional: boolean;
}